+++ src/app/modules/articles/articles.module.ts
@@ -5,7 +5,7 @@ import { ArticlesRoutingModule } from './articles-routing.module';
 import { ArticlesPageComponent } from './pages/articles-page/articles-page.component';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { ArticlePageComponent } from './pages/article-page/article-page.component';
-import { CreateArticleButtonComponent } from './admin/components/create-article-buttom/create-article-button.component';
+import { CreateArticleButtonComponent } from '../admin/articles-components/create-article-buttom/create-article-button.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AngularFireModule } from '@angular/fire/compat';
 import { AngularFirestoreModule } from '@angular/fire/compat/firestore';
